requirements.txt
asn1crypto==0.24.0
backports.ssl-match-hostname==3.5.0.1
bcrypt==3.1.7
cached-property==1.5.1
certifi==2017.7.27.1
cffi==1.12.3
chardet==3.0.4
Cheetah==2.4.4
Click==7.0
cloud-init==0.7.6
configobj==4.7.2
cryptography==2.7
decorator==3.4.0
docker==3.7.3
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.6
Flask==1.0.2
functools32==3.2.3.post2
gevent==1.3.7
greenlet==0.4.15
idna==2.6
iniparse==0.4
ipaddress==1.0.16
IPy==0.75
itsdangerous==1.1.0
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.12
jsonschema==2.6.0
kitchen==1.1.1
locustio==0.9.0
Markdown==2.6.9
MarkupSafe==1.0
meld3==0.6.10
msgpack==0.5.6
MySQL-python==1.2.5
oauth==1.0.1
paramiko==2.6.0
perf==0.1
pipenv==2018.11.26
policycoreutils-default-encoding==0.1
prettytable==0.7.2
pycparser==2.19
pycurl==7.19.0
pygobject==3.22.0
pygpgme==0.3
pyliblzma==0.5.3
PyNaCl==1.3.0
python-linux-procfs==0.4.9
pyudev==0.15
pyxattr==0.5.1
PyYAML==3.12
pyzmq==17.1.2
requests==2.18.4
schedutils==0.4
seobject==0.1
sepolicy==1.1
six==1.11.0
slip==0.4.0
slip.dbus==0.4.0
supervisor==3.1.4
texttable==0.9.1
typing==3.6.6
urlgrabber==3.10
urllib3==1.22
virtualenv==16.1.0
virtualenv-clone==0.4.0
websocket-client==0.56.0
Werkzeug==0.14.1
yum-metadata-parser==1.1.4